Revert "[ODS-5623] Make the Swagger Startup configurable (#1204)"
This reverts commit c4fab159a7fbe0282ee2be8f8cd75914dda3bf53.
