+++ Application/EdFi.Ods.SwaggerUI/Program.cs
@@ -5,11 +5,8 @@
 
 using System;
 using System.Threading.Tasks;
-using EdFi.Common.Extensions;
-using EdFi.Ods.Api.Helpers;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace EdFi.Ods.SwaggerUI
@@ -18,37 +15,21 @@ namespace EdFi.Ods.SwaggerUI
     {
         public static async Task Main(string[] args)
         {
-            string startupClassName;
+            await CreateHostBuilder(args).Build().RunAsync();
+        }
 
-            using (var hostBuilderInitial = Host.CreateDefaultBuilder(args).Build())
-            {
-                var config = hostBuilderInitial.Services.GetService<IConfiguration>();
-                startupClassName = config.GetValue<string>("StartupClassName");
-
-                AssemblyLoaderHelper.LoadAssembliesFromFolder(
-                    AssemblyLoaderHelper.GetPluginFolder(
-                        config.GetValue<string>("Plugin:Folder") ?? string.Empty));
-            }
-
-            var hostBuilder = Host.CreateDefaultBuilder(args)
+        public static IHostBuilder CreateHostBuilder(string[] args)
+            => Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(
                     webBuilder =>
                     {
                         webBuilder.ConfigureKestrel(
-                            serverOptions => { serverOptions.AddServerHeader = false; });
+                            serverOptions =>
+                            {
+                                serverOptions.AddServerHeader = false;
+                            });
 
-                        if (string.IsNullOrEmpty(startupClassName) || startupClassName.EqualsIgnoreCase("Startup"))
-                        {
-                            webBuilder.UseStartup<Startup>();
-                        }
-                        else
-                        {
-                            var startupFactory = new StartupFactory(startupClassName);
-                            webBuilder.UseStartup(context => startupFactory.CreateStartup(context));
-                        }
+                        webBuilder.UseStartup<Startup>();
                     });
-           
-            await hostBuilder.Build().RunAsync();
-        }
     }
 }
