--- Application/EdFi.Ods.SwaggerUI/SwaggerStartupBase.cs
@@ -1,145 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Licensed to the Ed-Fi Alliance under one or more agreements.
-// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
-// See the LICENSE and NOTICES files in the project root for more information.
-
-using System.Collections.Generic;
-using System.Reflection;
-using System.Text.Json;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpOverrides;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Swashbuckle.AspNetCore.SwaggerUI;
-
-namespace EdFi.Ods.SwaggerUI
-{
-    public abstract class SwaggerStartupBase
-    {
-        private readonly bool _useReverseProxyHeaders;
-        private readonly string _pathBase;
-        private readonly string _routePrefix;
-        private const string DefaultRoutePrefix = "swagger";
-
-        protected SwaggerStartupBase(IConfiguration configuration)
-        {
-            Configuration = configuration;
-            _useReverseProxyHeaders = Configuration.GetValue<bool>("UseReverseProxyHeaders");
-
-            var pathBase = Configuration.GetValue<string>("PathBase");
-            _routePrefix = Configuration.GetValue("SwaggerUIOptions:RoutePrefix", DefaultRoutePrefix);
-
-            if (!string.IsNullOrEmpty(pathBase))
-            {
-                _pathBase = "/" + pathBase.Trim('/');
-            }
-        }
-
-        private IConfiguration Configuration { get; }
-
-        public virtual void ConfigureServices(IServiceCollection services)
-        {
-            if (_useReverseProxyHeaders)
-            {
-                services.Configure<ForwardedHeadersOptions>(
-                    options =>
-                    {
-                        options.ForwardedHeaders = ForwardedHeaders.XForwardedFor
-                                                   | ForwardedHeaders.XForwardedHost
-                                                   | ForwardedHeaders.XForwardedProto;
-
-                        // Accept forwarded headers from any network and proxy
-                        options.KnownNetworks.Clear();
-                        options.KnownProxies.Clear();
-                    });
-            }
-
-            services.AddHealthChecks();
-        }
-
-        public virtual void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILogger<SwaggerStartupBase> logger)
-        {
-            string webApiUrl = Configuration.GetValue("WebApiVersionUrl", string.Empty);
-            var years = Configuration.GetSection("Years").Get<List<Years>>() ?? new List<Years>();
-            string swaggerStyleSheetPath = "../swagger.css";
-
-            if (!string.IsNullOrEmpty(_pathBase))
-            {
-                app.UsePathBase(_pathBase);
-            }
-
-            if (_useReverseProxyHeaders)
-            {
-                app.UseForwardedHeaders();
-            }
-
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-            }
-
-            app.Map(
-                "/appSettings.json", builder =>
-                {
-                    builder.Run(
-                        async context =>
-                        {
-                            context.Response.ContentType = "application/json";
-
-                            await context.Response.WriteAsync(
-                                JsonSerializer.Serialize(
-                                    new
-                                    {
-                                        WebApiVersionUrl = webApiUrl,
-                                        RoutePrefix = _routePrefix,
-                                        Years = years
-                                    }));
-                        });
-                });
-
-            logger.LogInformation($"WebApiUrl = '{webApiUrl}'");
-            logger.LogInformation($"UseReverseProxyHeaders = '{_useReverseProxyHeaders}'");
-            logger.LogInformation($"PathBase = '{_pathBase}'");
-            logger.LogInformation($"SwaggerStyleSheetPath = '{swaggerStyleSheetPath}'");
-
-            // routes for swagger: http://server/PATHBASE/ROUTEPREFIX/index.html
-            app.UseSwaggerUI(
-                options =>
-                {
-                    Configuration.Bind("SwaggerUIOptions", options);
-
-                    options.DocumentTitle = "Ed-Fi ODS API Documentation";
-
-                    options.OAuthScopeSeparator(" ");
-                    options.OAuthAppName(Assembly.GetExecutingAssembly().GetName().Name);
-                    options.DocExpansion(DocExpansion.None);
-                    options.DisplayRequestDuration();
-                    options.ShowExtensions();
-                    options.EnableValidator();
-                    options.InjectStylesheet(swaggerStyleSheetPath);
-
-                    options.IndexStream = ()
-                        => GetType().Assembly.GetManifestResourceStream("EdFi.Ods.SwaggerUI.Resources.Swashbuckle_index.html");
-
-                    options.ConfigObject.AdditionalItems["WebApiVersionUrl"] = webApiUrl;
-                    options.RoutePrefix = _routePrefix;
-                });
-
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
-            app.UseRouting();
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapHealthChecks("/health");
-            });
-        }
-    }
-}
