+++ Application/EdFi.Ods.SwaggerUI/appsettings.json
@@ -1,7 +1,6 @@
 {
   "WebApiVersionUrl": "",
   "SwaggerUIOptions": {
-    "StartupClassName": "",
     "OAuthConfigObject": {
       "ClientId": "",
       "ClientSecret": ""
